Module aws_sdk_chimesdkmessaging::error

Modules

bad_request_exception

See BadRequestException

conflict_exception

See ConflictException

forbidden_exception

See ForbiddenException

not_found_exception

See NotFoundException

resource_limit_exceeded_exception

See ResourceLimitExceededException

service_failure_exception

See ServiceFailureException

service_unavailable_exception

See ServiceUnavailableException

throttled_client_exception

See ThrottledClientException

unauthorized_client_exception

See UnauthorizedClientException

Structs

BadRequestException

The input parameters don't match the service's restrictions.

BatchCreateChannelMembershipError

ConflictException

The request could not be processed because of conflict in the current state of the

CreateChannelBanError

CreateChannelError

CreateChannelMembershipError

CreateChannelModeratorError

DeleteChannelBanError

DeleteChannelError

DeleteChannelMembershipError

DeleteChannelMessageError

DeleteChannelModeratorError

DescribeChannelBanError

DescribeChannelError

DescribeChannelMembershipError

DescribeChannelMembershipForAppInstanceUserError

DescribeChannelModeratedByAppInstanceUserError

DescribeChannelModeratorError

ForbiddenException

The client is permanently forbidden from making the request.

GetChannelMessageError

GetMessagingSessionEndpointError

ListChannelBansError

ListChannelMembershipsError

ListChannelMembershipsForAppInstanceUserError

ListChannelMessagesError

ListChannelModeratorsError

ListChannelsError

ListChannelsModeratedByAppInstanceUserError

NotFoundException

One or more of the resources in the request does not exist in the system.

RedactChannelMessageError

ResourceLimitExceededException

The request exceeds the resource limit.

SendChannelMessageError

ServiceFailureException

The service encountered an unexpected error.

ServiceUnavailableException

The service is currently unavailable.

ThrottledClientException

The client exceeded its request rate limit.

UnauthorizedClientException

The client is not currently authorized to make the request.

UpdateChannelError

UpdateChannelMessageError

UpdateChannelReadMarkerError

Enums

BatchCreateChannelMembershipErrorKind

CreateChannelBanErrorKind

CreateChannelErrorKind

CreateChannelMembershipErrorKind

CreateChannelModeratorErrorKind

DeleteChannelBanErrorKind

DeleteChannelErrorKind

DeleteChannelMembershipErrorKind

DeleteChannelMessageErrorKind

DeleteChannelModeratorErrorKind

DescribeChannelBanErrorKind

DescribeChannelErrorKind

DescribeChannelMembershipErrorKind

DescribeChannelMembershipForAppInstanceUserErrorKind

DescribeChannelModeratedByAppInstanceUserErrorKind

DescribeChannelModeratorErrorKind

GetChannelMessageErrorKind

GetMessagingSessionEndpointErrorKind

ListChannelBansErrorKind

ListChannelMembershipsErrorKind

ListChannelMembershipsForAppInstanceUserErrorKind

ListChannelMessagesErrorKind

ListChannelModeratorsErrorKind

ListChannelsErrorKind

ListChannelsModeratedByAppInstanceUserErrorKind

RedactChannelMessageErrorKind

SendChannelMessageErrorKind

UpdateChannelErrorKind

UpdateChannelMessageErrorKind

UpdateChannelReadMarkerErrorKind