Module aws_sdk_chimesdkmessaging::model

Modules

app_instance_user_membership_summary

See AppInstanceUserMembershipSummary

batch_channel_memberships

See BatchChannelMemberships

batch_create_channel_membership_error

See BatchCreateChannelMembershipError

channel

See Channel

channel_ban

See ChannelBan

channel_ban_summary

See ChannelBanSummary

channel_membership

See ChannelMembership

channel_membership_for_app_instance_user_summary

See ChannelMembershipForAppInstanceUserSummary

channel_membership_summary

See ChannelMembershipSummary

channel_message

See ChannelMessage

channel_message_summary

See ChannelMessageSummary

channel_moderated_by_app_instance_user_summary

See ChannelModeratedByAppInstanceUserSummary

channel_moderator

See ChannelModerator

channel_moderator_summary

See ChannelModeratorSummary

channel_summary

See ChannelSummary

identity

See Identity

messaging_session_endpoint

See MessagingSessionEndpoint

tag

See Tag

Structs

AppInstanceUserMembershipSummary

Summary of the membership details of an AppInstanceUser.

BatchChannelMemberships

The membership information, including member ARNs, the channel ARN, and membership

BatchCreateChannelMembershipError

A list of failed member ARNs, error codes, and error messages.

Channel

The details of a channel.

ChannelBan

The details of a channel ban.

ChannelBanSummary

Summary of the details of a ChannelBan.

ChannelMembership

The details of a channel member.

ChannelMembershipForAppInstanceUserSummary

Summary of the channel membership details of an AppInstanceUser.

ChannelMembershipSummary

Summary of the details of a ChannelMembership.

ChannelMessage

The details of a message in a channel.

ChannelMessageSummary

Summary of the messages in a Channel.

ChannelModeratedByAppInstanceUserSummary

Summary of the details of a moderated channel.

ChannelModerator

The details of a channel moderator.

ChannelModeratorSummary

Summary of the details of a ChannelModerator.

ChannelSummary

Summary of the details of a Channel.

Identity

The details of a user.

MessagingSessionEndpoint

The websocket endpoint used to connect to Amazon Chime SDK messaging.

Tag

Describes a tag applied to a resource.

Enums

ChannelMembershipType

ChannelMessagePersistenceType

ChannelMessageType

ChannelMode

ChannelPrivacy

ErrorCode

SortOrder