Module aws_sdk_chimesdkmessaging::operation

Structs

BatchCreateChannelMembership

Adds a specified number of users to a channel.

CreateChannel

Creates a channel to which you can add users and send messages.

CreateChannelBan

Permanently bans a member from a channel. Moderators can't add banned members to a

CreateChannelMembership

Adds a user to a channel. The InvitedBy response field is derived from the

CreateChannelModerator

Creates a new ChannelModerator. A channel moderator can:

DeleteChannel

Immediately makes a channel and its memberships inaccessible and marks them for

DeleteChannelBan

Removes a user from a channel's ban list.

DeleteChannelMembership

Removes a member from a channel.

DeleteChannelMessage

Deletes a channel message. Only admins can perform this action. Deletion makes messages

DeleteChannelModerator

Deletes a channel moderator.

DescribeChannel

Returns the full details of a channel in an Amazon Chime

DescribeChannelBan

Returns the full details of a channel ban.

DescribeChannelMembership

Returns the full details of a user's channel membership.

DescribeChannelMembershipForAppInstanceUser

Returns the details of a channel based on the membership of the specified

DescribeChannelModeratedByAppInstanceUser

Returns the full details of a channel moderated by the specified

DescribeChannelModerator

Returns the full details of a single ChannelModerator.

GetChannelMessage

Gets the full details of a channel message.

GetMessagingSessionEndpoint

The details of the endpoint for the messaging session.

ListChannelBans

Lists all the users banned from a particular channel.

ListChannelMemberships

Lists all channel memberships in a channel.

ListChannelMembershipsForAppInstanceUser

Lists all channels that a particular AppInstanceUser is a part of. Only an

ListChannelMessages

List all the messages in a channel. Returns a paginated list of

ListChannelModerators

Lists all the moderators for a channel.

ListChannels

Lists all Channels created under a single Chime App as a paginated list. You can specify

ListChannelsModeratedByAppInstanceUser

A list of the channels moderated by an AppInstanceUser.

RedactChannelMessage

Redacts message content, but not metadata. The message exists in the back end, but the

SendChannelMessage

Sends a message to a particular channel that the member is a part of.

UpdateChannel

Update a channel's attributes.

UpdateChannelMessage

Updates the content of a message.

UpdateChannelReadMarker

The details of the time when a user last read messages in a channel.